base: process: move content::SetProcessTitleFromCommandLine to //base
crrev.com/c/758962 introduced BrokerProcessPreSandboxHook() in
content/gpu/gpu_sandbox_hook_linux.cc because
SetProcessTitleFromCommandLine() was in service_manager package as the
comment in BrokerProcessPreSandboxHook() make an excuse.

However the PreSandboxHook is so complex that other broker processes
forget to call SetProcessTitleFromCommandLine().

By moving SetProcessTitleFromCommandLine() to //base package, the
complex PreSandboxHook is no longer needed (in the following CL).

Bug: b:328170541
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5348437
Commit-Queue: Shin Kawamura <kawasin@google.com>
Cr-Commit-Position: refs/heads/main@{#1270022}

+++ base/process/set_process_title_linux.cc
@@ -38,7 +38,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // this position within the glibc project, leaving applications caught in the
 // middle. (Also, only a very few applications need or want this anyway.)
 
-#include "content/common/set_process_title_linux.h"
+#include "base/process/set_process_title_linux.h"
 
 #include <stdarg.h>
 #include <stddef.h>
@@ -52,6 +52,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "base/files/file_util.h"
 #include "base/no_destructor.h"
+#include "base/numerics/safe_conversions.h"
 
 extern char** environ;
 
@@ -78,7 +79,8 @@ void setproctitle(const char* fmt, ...) {
     return;
 
   // The title can be up to the end of envp.
-  const size_t avail_size = g_envp_end - g_argv_start - 1;
+  const size_t avail_size =
+      base::checked_cast<size_t>(g_envp_end - g_argv_start - 1);
 
   // Linux 4.18--5.2 have a bug where we can never set a process title
   // shorter than the initial argv. Check if the bug exists in the current
@@ -104,11 +106,14 @@ void setproctitle(const char* fmt, ...) {
   size_t size;
   va_start(ap, fmt);
   if (fmt[0] == '-') {
-    size = vsnprintf(g_argv_start, avail_size, &fmt[1], ap);
+    size = base::checked_cast<size_t>(
+        vsnprintf(g_argv_start, avail_size, &fmt[1], ap));
   } else {
-    size = snprintf(g_argv_start, avail_size, "%s ", g_orig_argv0);
+    size = base::checked_cast<size_t>(
+        snprintf(g_argv_start, avail_size, "%s ", g_orig_argv0));
     if (size < avail_size)
-      size += vsnprintf(&g_argv_start[size], avail_size - size, fmt, ap);
+      size += base::checked_cast<size_t>(
+          vsnprintf(&g_argv_start[size], avail_size - size, fmt, ap));
   }
   va_end(ap);
 
@@ -123,7 +128,8 @@ void setproctitle(const char* fmt, ...) {
   // On buggy kernels we can never make the process title shorter than the
   // initial argv. In that case, just leave the remaining bytes filled with
   // null characters.
-  const size_t argv_size = g_argv_end - g_argv_start - 1;
+  const size_t argv_size =
+      base::checked_cast<size_t>(g_argv_end - g_argv_start - 1);
   if (!buggy_kernel && size < argv_size)
     g_argv_end[-1] = '.';
 }
